pyproject.toml
[project]
name = "autostream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "langchain-core>=1.3.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.9",
    "passlib>=1.7.4",
    "pydantic[email]>=2.13.3",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.45.0",
]
